Add files via upload
This algorithm computes a sparse transformation matrix needed in the facial reduction for our SDP relaxations. More precisely, it creates a sparse basis for the flow space of the bipartite representation of the directed graph G. The input of the algorithm is the m-by-n incidence matrix of a directed graph with n nodes and m arcs.